@@ -3,11 +3,11 @@ import { Route, Link } from 'react-router-dom'
 import PropTypes from 'prop-types'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { update } from '../../actions/account'
+// import { update } from '../../actions/account'
 // import { getAccessToken } from '../../reducers/account'
 
-import CSSModules from 'react-css-modules'
-import styles from './style.scss'
+// import CSSModules from 'react-css-modules'
+// import styles from './style.scss'
 
 import Shell from '../shell'
 
@@ -18,11 +18,12 @@ const Topic = ({ match }) => (
 )
 
 
-export default () => {
-  return (  <div>
-      <h3>22222</h3>
-    </div>)
-}
+// export default () => {
+//   return (  <div>
+//       <h3>22222</h3>
+//     </div>)
+// }
+
 
 // 纯组件
 export class Posts extends React.Component {
@@ -34,7 +35,7 @@ export class Posts extends React.Component {
   }
 
   componentDidMount() {
-    this.props.update('ttt')
+    // this.props.update('ttt')
     // console.log(this.props.location.params);
   }
 
@@ -44,7 +45,7 @@ export class Posts extends React.Component {
       <div>
 
       <div>
-        <h2 styleName="h2">Posts111</h2>
+        <h2>Posts111</h2>
         {/*
         <ul>
           <li>
@@ -78,23 +79,23 @@ export class Posts extends React.Component {
 
 }
 
-Posts = CSSModules(Posts, styles)
+// Posts = CSSModules(Posts, styles)
+//
+// Posts.propTypes = {
+//   update: PropTypes.func.isRequired
+// }
+//
+// const mapStateToProps = (state, props) => {
+//   return {
+//   }
+// }
+//
+// const mapDispatchToProps = (dispatch) => {
+//   return {
+//     update: bindActionCreators(update, dispatch)
+//   }
+// }
+//
+// Posts = connect(mapStateToProps,mapDispatchToProps)(Posts)
 
-Posts.propTypes = {
-  update: PropTypes.func.isRequired
-}
-
-const mapStateToProps = (state, props) => {
-  return {
-  }
-}
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    update: bindActionCreators(update, dispatch)
-  }
-}
-
-Posts = connect(mapStateToProps,mapDispatchToProps)(Posts)
-
-// export default Shell(Posts)
+export default Shell(Posts)
